Troubleshooting
I had to copy and paste my code into a different file explorer because my project was saved in a different place than where the Git Desktop was operating. This is not my first commit, but its the first one being processed. This has my map, basic structure for my app, but plots are missing